+++ src/lfx/src/lfx/services/settings/groups/__init__.py
@@ -1,0 +1,39 @@
+"""Logical groupings of Langflow settings.
+
+Each module defines a ``BaseModel`` mixin that owns a cohesive subset of fields
+plus their intra-group validators. They are composed into the final
+``Settings`` class in :mod:`lfx.services.settings.base`.
+
+Mixins inherit from ``BaseModel`` (not ``BaseSettings``) and are not intended
+to be instantiated directly.
+"""
+
+from lfx.services.settings.groups.cache import CacheSettings
+from lfx.services.settings.groups.components import ComponentsSettings
+from lfx.services.settings.groups.database import DatabaseSettings
+from lfx.services.settings.groups.mcp import McpSettings
+from lfx.services.settings.groups.observability import ObservabilitySettings
+from lfx.services.settings.groups.paths import PathSettings
+from lfx.services.settings.groups.runtime import RuntimeSettings
+from lfx.services.settings.groups.security import SecuritySettings
+from lfx.services.settings.groups.server import ServerSettings
+from lfx.services.settings.groups.storage import StorageSettings
+from lfx.services.settings.groups.telemetry import TelemetrySettings
+from lfx.services.settings.groups.ui import UiSettings
+from lfx.services.settings.groups.variables import VariablesSettings
+
+__all__ = [
+    "CacheSettings",
+    "ComponentsSettings",
+    "DatabaseSettings",
+    "McpSettings",
+    "ObservabilitySettings",
+    "PathSettings",
+    "RuntimeSettings",
+    "SecuritySettings",
+    "ServerSettings",
+    "StorageSettings",
+    "TelemetrySettings",
+    "UiSettings",
+    "VariablesSettings",
+]
